[#175798598] refactor and fix tests
