+++ src/__integrations__/pagopaApi.server.integration.test.ts
@@ -119,121 +119,4 @@ describe('Payment Manager Client', () => {
     expect(responseP.map((myRes: any) => myRes.status).getOrElse(404)).toEqual(200);
     await stubServerTerminator.terminate();
   });
-
-  it('should return successful response (200) when checkPayment endpoint is invoked with valid idPayment', async () => {
-    const HOST = process.env.PAYMENT_MANAGER_STUB_HOST as string;
-    const PORT = process.env.PAYMENT_MANAGER_STUB_PORT ? parseInt(process.env.PAYMENT_MANAGER_STUB_PORT, 10) : 5000;
-    const pmMockServer = pm.listen(PORT, HOST);
-    const stubServerTerminator = createHttpTerminator({ server: pmMockServer });
-
-    const paymentManagerClient = createClient({
-      baseUrl: `http://${HOST}:${PORT}`,
-      fetchApi: retryingFetch(fetch, 5000 as Millisecond, 5),
-    });
-
-    const idPayment = '8fa64d75-acb4-4a74-a87c-32f348a6a95f';
-
-    const response = await paymentManagerClient.checkPaymentUsingGET({
-      id: idPayment,
-    });
-
-    expect(response.isRight()).toBeTruthy();
-
-    expect(
-      response.fold(
-        _ => fail(),
-        response => response.status,
-      ),
-    ).toEqual(200);
-
-    expect(
-      response.fold(
-        _ => fail(),
-        response => response.value?.data?.idPayment,
-      ),
-    ).toEqual(idPayment);
-
-    await stubServerTerminator.terminate();
-  });
-
-  it('should return UnprocessableEntity error (422) when checkPayment endpoint is invoked with idPayment already retrived', async () => {
-    const HOST = process.env.PAYMENT_MANAGER_STUB_HOST as string;
-    const PORT = process.env.PAYMENT_MANAGER_STUB_PORT ? parseInt(process.env.PAYMENT_MANAGER_STUB_PORT, 10) : 5000;
-    const pmMockServer = pm.listen(PORT, HOST);
-
-    const stubServerTerminator = createHttpTerminator({ server: pmMockServer });
-    const paymentManagerClient = createClient({
-      baseUrl: `http://${HOST}:${PORT}`,
-      fetchApi: retryingFetch(fetch, 5000 as Millisecond, 5),
-    });
-
-    const idPayment = 'ca41570b-8c03-496b-9192-9284dec646d2';
-
-    expect(
-      (
-        await paymentManagerClient.checkPaymentUsingGET({
-          id: idPayment,
-        })
-      ).fold(
-        _ => fail(),
-        response => response.status,
-      ),
-    ).toEqual(422);
-
-    await stubServerTerminator.terminate();
-  });
-
-  it('should return not found error (404) when checkPayment endpoint is invoked with idPayment not found', async () => {
-    const HOST = process.env.PAYMENT_MANAGER_STUB_HOST as string;
-    const PORT = process.env.PAYMENT_MANAGER_STUB_PORT ? parseInt(process.env.PAYMENT_MANAGER_STUB_PORT, 10) : 5000;
-    const pmMockServer = pm.listen(PORT, HOST);
-
-    const stubServerTerminator = createHttpTerminator({ server: pmMockServer });
-    const paymentManagerClient = createClient({
-      baseUrl: `http://${HOST}:${PORT}`,
-      fetchApi: retryingFetch(fetch, 5000 as Millisecond, 5),
-    });
-
-    const idPayment = 'bn41570b-8c03-5432-9192-4444dec646d2';
-
-    expect(
-      (
-        await paymentManagerClient.checkPaymentUsingGET({
-          id: idPayment,
-        })
-      ).fold(
-        _ => fail(),
-        response => response.status,
-      ),
-    ).toEqual(404);
-
-    await stubServerTerminator.terminate();
-  });
-
-  it('should return generic error (500) when checkPayment endpoint fails with generic error', async () => {
-    const HOST = process.env.PAYMENT_MANAGER_STUB_HOST as string;
-    const PORT = process.env.PAYMENT_MANAGER_STUB_PORT ? parseInt(process.env.PAYMENT_MANAGER_STUB_PORT, 10) : 5000;
-    const pmMockServer = pm.listen(PORT, HOST);
-
-    const stubServerTerminator = createHttpTerminator({ server: pmMockServer });
-    const paymentManagerClient = createClient({
-      baseUrl: `http://${HOST}:${PORT}`,
-      fetchApi: retryingFetch(fetch, 5000 as Millisecond, 5),
-    });
-
-    const idPayment = 'ba41570b-77c03-496b-9192-9284dec646d2';
-
-    expect(
-      (
-        await paymentManagerClient.checkPaymentUsingGET({
-          id: idPayment,
-        })
-      ).fold(
-        _ => fail(),
-        response => response.status,
-      ),
-    ).toEqual(500);
-
-    await stubServerTerminator.terminate();
-  });
 });
